Apply clippy changes

+++ src/main.rs
@@ -77,8 +77,8 @@ fn size_scaling(windows: Res<Windows>, mut q: Query<(&Size, &mut Transform)>) {
     if let Some(window) = windows.get_primary() {
         for (sprite_size, mut transform) in q.iter_mut() {
             transform.scale = Vec3::new(
-                sprite_size.width / ARENA_WIDTH as f32 * window.width() as f32,
-                sprite_size.height / ARENA_HEIGHT as f32 * window.height() as f32,
+                sprite_size.width / ARENA_WIDTH as f32 * window.width(),
+                sprite_size.height / ARENA_HEIGHT as f32 * window.height(),
                 1.0,
             );
         }
@@ -93,8 +93,8 @@ fn position_translation(windows: Res<Windows>, mut q: Query<(&Position, &mut Tra
     if let Some(window) = windows.get_primary() {
         for (pos, mut transform) in q.iter_mut() {
             transform.translation = Vec3::new(
-                convert(pos.x as f32, window.width() as f32, ARENA_WIDTH as f32),
-                convert(pos.y as f32, window.height() as f32, ARENA_HEIGHT as f32),
+                convert(pos.x as f32, window.width(), ARENA_WIDTH as f32),
+                convert(pos.y as f32, window.height(), ARENA_HEIGHT as f32),
                 0.0,
             )
         }
